+++ src/components/RegistrationsIndex.vue
@@ -2,14 +2,13 @@
   <div class="hi">
     <p>
       Current Registrations coming soon
+      {{ this.registrations }}
     </p>
-    <p>Register for current season <router-link to="/user_registration">here</router-link></p>
-    <!-- <p>drag and drop to reorder</p> -->
   </div>
 </template>
 
 <script>
-import axios from 'axios'
+// import axios from 'axios'
 
 export default {
   name: 'RegistrationsIndex',
@@ -17,18 +16,12 @@ export default {
     registrations: Array,
     someProp: String
   },
+  data () {
+  },
   methods: {
-    async getRegistrations () {
-      // console.log('token', this.user.token)
-      await axios.get(process.env.VUE_APP_API_URL + '/registrations')
-        .then(resp =>
-          console.log('resp', resp)
-        )
-    }
   },
   mounted () {
     console.log('reg index mounted')
-    this.getRegistrations()
   }
 }
 </script>
