+++ src/views/RegistrationsView.vue
@@ -1,13 +1,17 @@
 <template>
   <div class="about">
     <h1>Current Registrations</h1>
-    <RegistrationsIndex />
+    <p>Register for current season <router-link to="/user_registration">here</router-link></p>
+    <RegistrationsIndex
+      :registrations="this.registrations"
+    />
     <!-- <p>{{ ApiUtilities }}</p> -->
   </div>
 </template>
 
 <script>
 // @ is an alias to /src
+import axios from 'axios'
 import RegistrationsIndex from '@/components/RegistrationsIndex.vue'
 import { ApiUtilities } from '../../utils/ApiUtilities.js'
 
@@ -16,12 +20,27 @@ export default {
   components: {
     RegistrationsIndex
   },
-  methods () {
+  data () {
+    return {
+      registrations: [],
+      um: 'hi'
+    }
+  },
+  methods: {
     // methods
+    async getRegistrations () {
+      // console.log('token', this.user.token)
+      await axios.get(process.env.VUE_APP_API_URL + '/user_registrations/1')
+        .then(resp =>
+        // { this.registrations = resp.data }
+          console.log('resp', resp)
+        )
+    }
   },
   mounted () {
     console.log('apiUtilities', ApiUtilities)
     // console.log('reg index mounted')
+    this.getRegistrations()
   }
 }
 </script>
